# blur performance improved