blur 2560x1440 40 radius:

before
image: `blur ............................ 5970.127 ms   5970.127 ms    ±0.000     x1`
mask: `blur ............................. 569.820 ms    574.004 ms    ±2.479     x9`

after
image: `blur ............................ 4186.500 ms   4284.556 ms  ±138.673     x2`
mask: `blur ............................. 387.028 ms    393.635 ms    ±4.447    x13`